trajectory refactor work
