+++ core/src/visad/java3d/ShadowFunctionOrSetTypeJ3D.java
@@ -32,14 +32,12 @@ import visad.util.ThreadManager;
 import javax.media.j3d.*;
 
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Vector;
 
 import java.rmi.*;
 
 import java.awt.image.*;
-import java.util.Arrays;
 
 
 /**
@@ -81,9 +79,9 @@ public class ShadowFunctionOrSetTypeJ3D extends ShadowTypeJ3D {
   float[] intrpU;
   float[] intrpV;
   float[] intrpW;
-  Interpolation uInterp;
-  Interpolation vInterp;
-  Interpolation wInterp;
+  Interpolator uInterp;
+  Interpolator vInterp;
+  Interpolator wInterp;
   float[][] values0;
   float[][] values1;
   float[][] values2;
@@ -272,7 +270,7 @@ public class ShadowFunctionOrSetTypeJ3D extends ShadowTypeJ3D {
         addSwitch(group, swit, control, domainSet, renderer);
       }
       else {
-        double[] times = Trajectory.getTimes((Gridded1DSet)anim1DdomainSet);
+        double[] times = TrajectoryManager.getTimes((Gridded1DSet)anim1DdomainSet);
         java.util.Arrays.sort(times);
         int len = times.length;
         double avgTimeStep = (times[len-1] - times[0])/(len-1);
@@ -1421,7 +1419,23 @@ System.out.println("Texture.BASE_LEVEL_LINEAR = " + Texture.BASE_LEVEL_LINEAR); 
        e.printStackTrace();
     }
   }
+  
+  public static BranchGroup addToDetachGroup(ShadowType shadow, VisADGeometryArray trcrArray, ArrayList<float[]> achrArrays, GraphicsModeControl mode, float constant_alpha, float[] constant_color) throws VisADException {
+     BranchGroup branch = (BranchGroup) shadow.makeBranch();
+     branch.setCapability(BranchGroup.ALLOW_CHILDREN_EXTEND);
+     branch.setCapability(BranchGroup.ALLOW_CHILDREN_WRITE);
 
+     /* This branch is detached from 'branch' during auto resizing. 
+        New resized trcrArray is then added back to 'branch' */
+     BranchGroup trcrBranch = (BranchGroup) shadow.makeBranch();
+
+     shadow.addToGroup(trcrBranch, trcrArray, mode, constant_alpha, constant_color);
+     branch.addChild(trcrBranch);
+
+     return branch;
+  }
+
+  
   /** render accumulated Vector of value_array-s to
       and add to group; then clear AccumulationVector */
   public void postProcess(Object group) throws VisADException {
@@ -1450,56 +1464,32 @@ System.out.println("Texture.BASE_LEVEL_LINEAR = " + Texture.BASE_LEVEL_LINEAR); 
   }
 
   private void doTrajectory() throws VisADException, RemoteException {
+    ArrayList<FlowInfo> flowInfoList = Range.getAdaptedShadowType().getFlowInfo();
+    dataDomainLength = anim1DdomainSet.getLength();
+    
     RendererJ3D renderer = (RendererJ3D) getLink().getRenderer();
     ProjectionControl pCntrl = renderer.getDisplay().getProjectionControl();
     FixedSizeListener listener = null;
+    
+    TrajectoryManager trajMan = new TrajectoryManager(renderer, trajParams, flowInfoList, dataDomainLength);
     
     trcrEnabled = trcrEnabled && (trajForm == LINE);    
     
     if (autoSizeTrcr && trcrEnabled) {
       listener = new FixedSizeListener(pCntrl, this);
-      Trajectory.setListener(pCntrl, listener, flowCntrl);
+      trajMan.setListener(pCntrl, listener, flowCntrl);
       listener.lock();
     }
-    double scale = Trajectory.getScaleX(pCntrl); // current dispaly scale
+    double scale = TrajectoryManager.getScaleX(pCntrl); // current dispaly scale
     
-    initTrajectory(renderer);
-    Trajectory.initCleanUp(flowMap, flowCntrl, pCntrl, renderer.getDisplay());
+    trajMan.initCleanUp(flowMap, flowCntrl, pCntrl, display);
     
     double trcrSizeRatio = 1;
-    TrajectoryParams lastTrajParams = Trajectory.getLastTrajParams(flowCntrl, trajParams);
-    if (lastTrajParams != null) {
-      canUseTrajCache = lastTrajParams.equals(trajParams);
-      trcrSizeRatio = trajParams.getMarkerSize()/lastTrajParams.getMarkerSize();
-    }
-    
-    Object trajCache = null;
-    if (trajCachingEnabled) {
-       trajCache = Trajectory.setCache(flowCntrl, canUseTrajCache);
-    }
-    else {
-       canUseTrajCache = false;
-    }
-    
-    intrpU = new float[numSpatialPts];
-    intrpV = new float[numSpatialPts];
-    intrpW = new float[numSpatialPts];
 
-    uInterp = new Interpolation(trajDoIntrp, numSpatialPts);
-    vInterp = new Interpolation(trajDoIntrp, numSpatialPts);
-    wInterp = new Interpolation(trajDoIntrp, numSpatialPts);
-
-    values0 = null;
-    values1 = null;
-    values2 = null;
-    values3 = null;
-    values0_last = null;
-
-    double[] times = Trajectory.getTimes((Gridded1DSet)anim1DdomainSet);
-    double[] timeSteps = Trajectory.getTimeSteps((Gridded1DSet)anim1DdomainSet);
+    double[] times = TrajectoryManager.getTimes((Gridded1DSet)anim1DdomainSet);
+    double[] timeSteps = TrajectoryManager.getTimeSteps((Gridded1DSet)anim1DdomainSet);
     timeAccum = 0;
 
-    ArrayList<FlowInfo> flowInfoList = Range.getAdaptedShadowType().getFlowInfo();
     VisADGeometryArray[] auxArray = new VisADGeometryArray[1];
     
     for (int k=0; k<dataDomainLength-1; k++) {
@@ -1507,31 +1497,11 @@ System.out.println("Texture.BASE_LEVEL_LINEAR = " + Texture.BASE_LEVEL_LINEAR); 
       
       FlowInfo info = flowInfoList.get(i);
       
-      if (!canUseTrajCache) {
-         array = computeTrajectories(k, flowInfoList, times, timeSteps, auxArray);
-         achrArrays = new ArrayList<float[]>();
-         if (trajectories.size() > 0) {
-            trcrArray = Trajectory.makeTracerGeometry(trajectories, trcrArrays, achrArrays, direction, trcrSize, dspScale, true);   
-            if (trajCachingEnabled) {
-              Trajectory.cacheTrajArray(trajCache, array);
-              Trajectory.cacheArray(trajCache, auxArray[0]);
-              Trajectory.cacheTrcrArray(trajCache, trcrArray, achrArrays);
-            }
-            trcrArray = Trajectory.scaleGeometry(trcrArray, achrArrays, (float)(1.0/scale));      
-         }
-      }
-      else  {
-         array = Trajectory.getCachedTraj(trajCache, k);
-         auxArray[0] = Trajectory.getCachedArray(trajCache, k);
-         trcrArray = Trajectory.getCachedTrcr(trajCache, k);
-         achrArrays = Trajectory.getCachedAncr(trajCache, k);
-         if (trcrSizeRatio != 1.0) {
-           trcrArray = Trajectory.scaleGeometry(trcrArray, achrArrays, (float)trcrSizeRatio);            
-           Trajectory.cacheTrcrArray(trajCache, k, trcrArray, achrArrays);
-         }
-         if (autoSizeTrcr) {
-           trcrArray = Trajectory.scaleGeometry(trcrArray, achrArrays, (float)(1.0/scale));
-         }
+      array = trajMan.computeTrajectories(k, timeAccum, times, timeSteps, auxArray);
+      achrArrays = new ArrayList<float[]>();
+      if (trajMan.getNumberOfTrajectories() > 0) {
+        trcrArray = trajMan.makeTracerGeometry(trcrArrays, achrArrays, direction, trcrSize, dspScale, true);
+        trcrArray = TrajectoryManager.scaleGeometry(trcrArray, achrArrays, (float)(1.0/scale));      
       }
       
       GraphicsModeControl mode = (GraphicsModeControl) info.mode.clone();
@@ -1555,7 +1525,7 @@ System.out.println("Texture.BASE_LEVEL_LINEAR = " + Texture.BASE_LEVEL_LINEAR); 
       final BranchGroup node = (BranchGroup) swit.getChild(i);
           
       if (trcrEnabled) {
-        BranchGroup trcrBG = Trajectory.makeTracerBranch(this, trcrArray, achrArrays, mode, info.constant_alpha, info.constant_color);
+        BranchGroup trcrBG = addToDetachGroup(this, trcrArray, achrArrays, mode, info.constant_alpha, info.constant_color);
         ((BranchGroup)switB.getChild(i)).addChild(trcrBG);
         if (listener != null) {
           listener.add(trcrBG, trcrArray, achrArrays, mode, info.constant_alpha, info.constant_color);
@@ -1579,7 +1549,7 @@ System.out.println("Texture.BASE_LEVEL_LINEAR = " + Texture.BASE_LEVEL_LINEAR); 
       GraphicsModeControl mode = (GraphicsModeControl) finfo.mode.clone();
 
       if (trcrEnabled) {
-        BranchGroup trcrBG = Trajectory.makeTracerBranch(this, trcrArray, achrArrays, mode, finfo.constant_alpha, finfo.constant_color);
+        BranchGroup trcrBG = addToDetachGroup(this, trcrArray, achrArrays, mode, finfo.constant_alpha, finfo.constant_color);
         ((BranchGroup)switB.getChild(idx)).addChild(trcrBG);
         if (listener != null) {
           listener.add(trcrBG, trcrArray, achrArrays, mode, finfo.constant_alpha, finfo.constant_color);
@@ -1594,208 +1564,7 @@ System.out.println("Texture.BASE_LEVEL_LINEAR = " + Texture.BASE_LEVEL_LINEAR); 
        }
     }
   }
-  
-     private void initTrajectory(DataRenderer renderer) throws VisADException, RemoteException {
-         /* Get start points, use first spatial_set locs for now. 
-            Eventually want to include a time for start */
-         ArrayList<FlowInfo> flowInfoList = Range.getAdaptedShadowType().getFlowInfo();
-         FlowInfo info = flowInfoList.get(0);
-         Gridded3DSet spatial_set0 = (Gridded3DSet) info.spatial_set;
-         GriddedSet spatialSetTraj = Trajectory.makeSpatialSetTraj(spatial_set0);
-
-         byte[][] color_values = info.color_values;
-         if (info.trajColors != null) color_values = info.trajColors;
-         clrDim = color_values.length;
-
-         numSpatialPts = spatial_set0.getLength();
-         Trajectory.markGrid = new boolean[numSpatialPts];
-         Trajectory.markGridTime = new int[numSpatialPts];
-         java.util.Arrays.fill(Trajectory.markGrid, false);
-         java.util.Arrays.fill(Trajectory.markGridTime, 0);
-
-         startClrs = new byte[clrDim][];
-         if (startPts == null) { //get from domain set
-           float[][] vec;
-           if (true) {
-              float[][] flowVals = Trajectory.convertFlowUnit(info.flow_values, info.flow_units);
-              vec = ShadowType.adjustFlowToEarth(info.which, flowVals, spatial_set0.getSamples(false), 1f, renderer);
-           }
-           startPts = new float[3][];
-           Trajectory.getStartPointsFromDomain(trajForm, trajSkip, zStart, zSkip, spatial_set0, color_values, startPts, startClrs, vec, ribbonWidthFac);
-         }
-         else {
-           if (!startPointType.equals(Display.DisplaySpatialCartesianTuple)) {
-             throw new VisADException("startPointType must be convertable"
-                     + " with (Display.XAxis, Display.YAxis, Display.ZAxis) ");
-           }
-           /* TODO: assuming earth navigated display coordinate system*/
-           /*
-           CoordinateSystem dspCoordSys = getLink().getRenderer().getDisplayCoordinateSystem();
-           float[][] fltVals = new float[startPts.length][startPts[0].length];
-           for (int i=0; i<fltVals.length; i++) {
-             System.arraycopy(startPts[i], 0, fltVals[i], 0, fltVals[i].length);
-           }
-           startPts = dspCoordSys.toReference(fltVals);
-           */
-
-           int[] clrIdxs;
-           if (spatialSetTraj.getManifoldDimension() == 2) {
-               clrIdxs = spatialSetTraj.valueToIndex(new float[][] {startPts[0], startPts[1]});
-           } else {
-               clrIdxs = spatialSetTraj.valueToIndex(startPts);
-           }
-           int num = clrIdxs.length;
-           startClrs[0] = new byte[num];
-           startClrs[1] = new byte[num];
-           startClrs[2] = new byte[num];
-           if (clrDim == 4) startClrs[3] = new byte[num];
-           for (int i=0; i<num; i++) {
-             int clrIdx = clrIdxs[i];
-             if (clrIdx < 0) continue;
-             startClrs[0][i] = color_values[0][clrIdx];
-             startClrs[1][i] = color_values[1][clrIdx];
-             startClrs[2][i] = color_values[2][clrIdx];
-             if (clrDim == 4) startClrs[3][i] = color_values[3][clrIdx];
-           }
-         }
-     }
-     
-     private VisADGeometryArray computeTrajectories(int k, ArrayList<FlowInfo> flowInfoList, double[] times, double[] timeSteps, VisADGeometryArray[] auxArray) throws VisADException, RemoteException {
-          int i = (direction < 0) ? ((dataDomainLength-1) - k) : k;
-
-          FlowInfo info = flowInfoList.get(i);
-          byte[][] color_values = info.color_values;
-          Gridded3DSet spatial_set = (Gridded3DSet) info.spatial_set;
-          GriddedSet spatialSetTraj = Trajectory.makeSpatialSetTraj(spatial_set);
-          
-          if (!manualIntrpPts && trajDoIntrp) {
-            numIntrpPts = Trajectory.getNumIntrpPts(info, 50f, timeSteps[i]);
-          }
-          
-          float timeStep = (float) timeSteps[i]/numIntrpPts;
-
-          if ((k==0) || (timeAccum >= trajRefreshInterval)) { // for non steady state trajectories (refresh frequency)
-             trajectories = new ArrayList<Trajectory>();
-             java.util.Arrays.fill(Trajectory.markGrid, false);
-             if (direction > 0) {
-               Trajectory.makeTrajectories(direction*times[i], trajectories, startPts, startClrs, spatialSetTraj);
-             }
-             else { //TODO: make this work eventually
-               //Trajectory.makeTrajectories(direction*times[i], trajectories, trajSkip, color_values, setLocs, lens);
-             }
-          }
-
-          // commented out when not using markGrid logic for starting/ending trajs
-          //Trajectory.makeTrajectories(times[i], trajectories, 6, color_values, setLocs, lens);
-          /*
-          Trajectory.checkTime(i); // for steady-state only
-          if ((i % 4) == 0) { // use for steady-state wind field
-            Trajectory.makeTrajectories(direction*times[i], trajectories, trajSkip, color_values, setLocs, lens);
-          }
-          */
-
-          double x0 = (double) direction*i;
-          double x1 = (double) direction*(i+direction*1);
-          double x2 = (double) direction*(i+direction*2);
-
-
-          // Even time steps: access fields, update interpolator and compute.
-          // Odd time steps: just compute for second half of 3 point (2 gap) interval.
-          if ((k % 2) == 0) {
-            FlowInfo flwInfo;
-
-            if (values0 == null) {
-              flwInfo = flowInfoList.get(i);
-              values0 = Trajectory.convertFlowUnit(flwInfo.flow_values, flwInfo.flow_units);
-            }
-
-            if (values1 == null) {
-              flwInfo = flowInfoList.get(i+direction*1);
-              values1 = Trajectory.convertFlowUnit(flwInfo.flow_values, flwInfo.flow_units);
-            }
-
-            // make sure we don't access more data than we have, but keep iterating and 
-            // computing to the end to use the data we've already pulled in.
-            if (k < dataDomainLength-3) {
-                flwInfo = flowInfoList.get(i+direction*2);
-                values2 = Trajectory.convertFlowUnit(flwInfo.flow_values, flwInfo.flow_units);
-
-                // smoothing done here -----------------------
-                flwInfo = flowInfoList.get(i+direction*3);
-                values3 = Trajectory.convertFlowUnit(flwInfo.flow_values, flwInfo.flow_units);
-
-                if (values0_last != null) {
-                  values0 = Trajectory.smooth(values0_last, values0, values1, smoothParams);
-                }
-                values1 = Trajectory.smooth(values0, values1, values2, smoothParams);
-                values2 = Trajectory.smooth(values1, values2, values3, smoothParams);
-                // ------- end smoothing
-
-                // update interpolator 
-                uInterp.next(x0, x1, x2, values0[0], values1[0], values2[0]);
-                vInterp.next(x0, x1, x2, values0[1], values1[1], values2[1]);
-                wInterp.next(x0, x1, x2, values0[2], values1[2], values2[2]);
-            }
-            
-            if (k == dataDomainLength-3) { // make sure we smoothly handle the last three time steps
-                uInterp.next(x0, x1, x2, values0[0], values1[0], values2[0]);
-                vInterp.next(x0, x1, x2, values0[1], values1[1], values2[1]);
-                wInterp.next(x0, x1, x2, values0[2], values1[2], values2[2]);
-            }
-          }
-          else {
-            values0_last = values1;
-            values0 = values2;
-            values1 = values3;
-            
-            if (k == dataDomainLength-3) { // make sure we smootly handle the last three time steps
-                uInterp.next(x0, x1, x2, values0[0], values1[0], values2[0]);
-                vInterp.next(x0, x1, x2, values0[1], values1[1], values2[1]);
-                wInterp.next(x0, x1, x2, values0[2], values1[2], values2[2]);
-            }
-          }
-
-          int numTrajectories = trajectories.size();
-          Trajectory.reset(trajectories, numTrajectories*numIntrpPts, clrDim);
-
-          for (int ti=0; ti<numIntrpPts; ti++) { // additional points per domain time step
-            double dst = (x1 - x0)/numIntrpPts;
-            double xt = x0 + dst*ti;
-             
-            Trajectory.updateInterpolators(trajectories, numSpatialPts, uInterp, vInterp, wInterp);
-            
-            uInterp.interpolate(xt, intrpU);
-            vInterp.interpolate(xt, intrpV);
-            wInterp.interpolate(xt, intrpW);
-            
-            for (int t=0; t<numTrajectories; t++) {
-              Trajectory traj = trajectories.get(t);
-              traj.currentTimeIndex = direction*i;
-              traj.currentTime = direction*times[i];
-              traj.forward(info, new float[][] {intrpU, intrpV, intrpW}, color_values, spatialSetTraj, direction, timeStep);
-            }
-
-          } // inner time loop (time interpolation)
-
-          switch (trajForm) {
-            case 0:
-              array = Trajectory.makeGeometry();
-              break;
-            case 1:
-              array = Trajectory.makeFixedWidthRibbon(trajectories, ribbonWidthFac);
-              break;
-            case 2:
-              array = Trajectory.makeCylinder(trajectories, auxArray, cylWidth);
-              break;
-            case 3:
-              array = Trajectory.makeDeformableRibbon(trajectories);
-              break;
-          }
-          
-          trajectories = Trajectory.clean(trajForm, trajectories, trajLifetime);
-          return array;
-     }
-  }
+}
 
 
   class SwitchListener extends Switch {
