+++ core/src/visad/util/CubicInterpolator.java
@@ -1,12 +1,12 @@
-package visad.java3d;
+package visad.util;
 
 import Jama.Matrix;
 import Jama.LUDecomposition;
 import java.util.Arrays;
 
-public class Interpolation {
+public class CubicInterpolator {
 
-      Jama.LUDecomposition solver;
+      LUDecomposition solver;
 
       double[][] solution = null;
 
@@ -25,10 +25,7 @@ public class Interpolation {
       boolean[] needed = null;
       boolean[] computed = null;
       
-      public Interpolation() {
-      }
-
-      public Interpolation(boolean doIntrp, int numSpatialPts) {
+      public CubicInterpolator(boolean doIntrp, int numSpatialPts) {
          this.doIntrp = doIntrp;
          this.numSpatialPts = numSpatialPts;
          this.solution = new double[4][numSpatialPts];
@@ -38,7 +35,7 @@ public class Interpolation {
          Arrays.fill(computed, false);
       }
 
-      void buildSolver() {
+      private void buildSolver() {
          double x0_p3 = x0*x0*x0;
          double x1_p3 = x1*x1*x1;
          double x2_p3 = x2*x2*x2;
@@ -53,7 +50,7 @@ public class Interpolation {
                 {x2_p3, x2_p2, x2, 1},
                 {3*x0_p2, 2*x0, 1, 0}}, 4, 4);
 
-         solver = new Jama.LUDecomposition(coeffs);
+         solver = new LUDecomposition(coeffs);
       }
 
       public void interpolate(double xt, float[] interpValues) {
@@ -106,7 +103,7 @@ public class Interpolation {
           }
       }
       
-      public void getSolution() {
+      private void getSolution() {
          for (int k=0; k<numSpatialPts; k++) {
             if (!this.needed[k]) {
                 continue;
@@ -115,30 +112,23 @@ public class Interpolation {
             double y1 = values1[k];
             double y2 = values2[k];
 
-            // TODO: for now always initialize first derivative at first point with estimate from
-            // the first two data pts instead of using derivative from cubic polynomial fit at the
-            // last point.  This works pretty well, but can be improved. So set this to "true".
-            if (true) {
-               double D1 = (y1 - y0)/(x1 - x0);
-               double[] sol = getSolution(y0, y1, y2, D1);
-               solution[0][k] = sol[0];
-               solution[1][k] = sol[1];
-               solution[2][k] = sol[2];
-               solution[3][k] = sol[3];
-            }
-            else {
-               double D1 = cubic_poly_D1(x0, solution[0][k], solution[1][k], solution[2][k]);
-               double[] sol = getSolution(y0, y1, y2, D1);
-               solution[0][k] = sol[0];
-               solution[1][k] = sol[1];
-               solution[2][k] = sol[2];
-               solution[3][k] = sol[3];
-            }
+            // TODO: Initialize first derivative at first point with estimate from the
+            // first two data pts instead of using derivative from cubic polynomial fit
+            // at the last point.  This works pretty well, but can be improved.
+            double D1 = (y1 - y0)/(x1 - x0);
+            //double D1 = cubic_poly_D1(x0, solution[0][k], solution[1][k], solution[2][k]);     
+            
+            double[] sol = getSolution(y0, y1, y2, D1);
+            solution[0][k] = sol[0];
+            solution[1][k] = sol[1];
+            solution[2][k] = sol[2];
+            solution[3][k] = sol[3];
+            
             computed[k] = true;
          }
       }
 
-      public double[] getSolution(double y0, double y1, double y2, double D1) {
+      private double[] getSolution(double y0, double y1, double y2, double D1) {
         Matrix constants = new Matrix(new double[][]
              { {y0}, {y1}, {y2}, {D1} }, 4, 1);
